+++ routes/web.php
@@ -18,23 +18,23 @@ use Statamic\Facades\Site;
 // Dynamic Token route for posting a form with Ajax.
 Route::get('/!/DynamicToken/refresh', [DynamicToken::class, 'getRefresh']);
 
-// The Sitemap route to the sitemap.xml
+// The Sitemap Index route for listing sitemaps of all (multi)sites.
+Route::statamic('/sitemaps.xml', 'sitemap/sitemaps', [
+    'layout' => null,
+    'content_type' => 'application/xml'
+]);
+
+// The Default Site Sitemap route.
 Route::statamic('/sitemap.xml', 'sitemap/sitemap', [
     'layout' => null,
     'content_type' => 'application/xml'
 ]);
 
-// The Sitemap route to the sitemap.xml for multisites
-// Route::statamic('/{site_handle}/sitemap.xml', 'sitemap/sitemap', [
-//     'layout' => null,
-//     'content_type' => 'application/xml'
-// ]);
-
-// The Sitemap Index route for listing sitemaps of multisites
-// Route::statamic('/sitemaps.xml', 'sitemap/sitemap_index', [
-//     'layout' => null,
-//     'content_type' => 'application/xml'
-// ]);
+// The Multisite Site Sitemap route(s).
+Route::statamic('/{site_handle}/sitemap.xml', 'sitemap/sitemap', [
+    'layout' => null,
+    'content_type' => 'application/xml'
+]);
 
 // The Manifest route to the manifest.json
 Route::statamic('/site.webmanifest', 'manifest/manifest', [